DataIngestion/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.0
pydantic==2.12.5
sqlalchemy==2.0.22
asyncpg==0.27.0
aiokafka==0.10.0
python-dateutil==2.8.2
ujson==5.10.0
loguru==0.7.0
